replace const.py with the const submodule
