+++ src/gpo/const/__init__.py
@@ -1,0 +1,5 @@
+"""A GPO submodule for utilities and constants"""
+import five.src.gpo.const.fields
+import five.src.gpo.const.option
+import five.src.gpo.const.error
+import five.src.gpo.const.util
